+++ src/interscale/model/__init__.py
@@ -1,6 +1,6 @@
 from .base._base_model import BaseModel
-from .CombinedModel import CombinedModel
-from .GlobalModel import GlobalModel
-from .LocalModel import LocalModel
+from .combined_model import CombinedModel
+from .global_model import GlobalModel
+from .local_model import LocalModel
 
 __all__ = ["BaseModel", "LocalModel", "GlobalModel", "CombinedModel"]
