+++ src/interscale/model/combined_model.py
