+++ src/interscale/model/global_model.py
