+++ src/interscale/model/local_model.py
